real time body temp

+++ api_smartbracelet/controllers.js
@@ -2,7 +2,6 @@ const prisma = require("../prisma/client");
 const webpush = require("web-push");
 const TEMPRATURE_RULE_1 = 38; //Jika lebih dari 38 beri reminder untuk minum obat
 const TEMPRATURE_RULE_2 = 41; //Jika lebih dari 41 beri notifikasi untuk pergi ke dokter
-
 exports.generateId = async (req, res) => {
     try {
         const characters =
@@ -40,6 +39,7 @@ exports.generateId = async (req, res) => {
 
 exports.updateTemprature = async (data, feedback) => {
     const body = JSON.parse(data);
+    feedback.socket.emit(`/temp/${body.id}`, body.temp);
     try {
         const bracelet = await prisma.smartBracelet.update({
             where: {
@@ -62,49 +62,54 @@ exports.updateTemprature = async (data, feedback) => {
         });
 
         let notificiationData = {};
+        let showNotif = false;
         const userId = bracelet.SmartMedicine.User.id;
         if (body.temp >= TEMPRATURE_RULE_1 && body.temp < TEMPRATURE_RULE_2) {
+            showNotif = true;
             notificiationData["title"] = "Body temperature warning";
             notificiationData["body"] =
                 "Are you feel sick! Your body temperature is high. Take immediate action to cool down and consider taking appropriate medication. Prioritize your health and well-being.";
         }
 
         if (body.temp > TEMPRATURE_RULE_2) {
+            showNotif = true;
             notificiationData["title"] = "Body temperature warning";
             notificiationData["body"] =
                 "Are you feel sick! Your body temperature is high. It is strongly advised to consult a doctor promptly for a thorough evaluation and appropriate medical guidance";
         }
 
-        const payload = JSON.stringify(notificiationData);
-        const user = await prisma.user.findUnique({
-            where: { id: userId },
-            select: {
-                subscription: {
-                    select: {
-                        subscriptionExpiredAt: true,
-                        identifier: true,
-                        subscriptionToken: true,
+        if (showNotif) {
+            const payload = JSON.stringify(notificiationData);
+            const user = await prisma.user.findUnique({
+                where: { id: userId },
+                select: {
+                    subscription: {
+                        select: {
+                            subscriptionExpiredAt: true,
+                            identifier: true,
+                            subscriptionToken: true,
+                        },
                     },
                 },
-            },
-        });
+            });
 
-        user.subscription.forEach(async (subs) => {
-            // Hanya user yang masih login yang bisa menerima notifikasi
-            if (new Date() < new Date(subs.subscriptionExpiredAt)) {
-                const subscription = JSON.parse(subs.subscriptionToken);
-                console.log("Notification Wes Send");
-                webpush
-                    .sendNotification(subscription, payload)
-                    .catch((err) => console.error(`ERR: ${err}`));
-            } else {
-                // Jika Token Sudah Kadaluarsa maka hapus dari database
-                console.log("User Subscription Expired");
-                await prisma.subscription.delete({
-                    where: { identifier: subs.identifier },
-                });
-            }
-        });
+            user.subscription.forEach(async (subs) => {
+                // Hanya user yang masih login yang bisa menerima notifikasi
+                if (new Date() < new Date(subs.subscriptionExpiredAt)) {
+                    const subscription = JSON.parse(subs.subscriptionToken);
+                    console.log("Notification Wes Send");
+                    webpush
+                        .sendNotification(subscription, payload)
+                        .catch((err) => console.error(`ERR: ${err}`));
+                } else {
+                    // Jika Token Sudah Kadaluarsa maka hapus dari database
+                    console.log("User Subscription Expired");
+                    await prisma.subscription.delete({
+                        where: { identifier: subs.identifier },
+                    });
+                }
+            });
+        }
 
         console.log("Success Update Temprature");
     } catch (error) {
